OPENGL3 - 120 FPS - 3-07-26
Update 3-07-26
- Atualização do OpenGL DE 1.0 para 3.3 (BETA)
- Importado algumas renderizações para a GPU (BETA)
- Desbloqueado 120 FPS rodando pela GPU (BETA)
- Melhorado todo sistema de FPS rodando pela GPU (BETA)
- Inserido Sistema de Shaders e VBO para melhor qualidade gráfica e melhor Desempenho
- Inserido controle de Volume da Musica separado do Sons do jogo, agora e possivel controlar o volume das músicas
- Novo Arquivo no MainInfo: UpdateColors.txt para definir as cores do sistema e update e dicas que aparecem na tela
- Novo sistema de Ranking F8 (BETA) com renderização dos personagens e max 20 Resultados (execute a query CustomRanking_F8.sql na pasta Querys do MuServer)
- Novo sistema de CustomAttackAutoResume (execute a query: CustomAttack_OffAttack.sql na pasta Querys do MuServer)
- Novo sistema de CustomOffAttackAutoResume (execute a query: CustomAttack_OffAttack.sql na pasta Querys do MuServer)
- Fixado não remover o Vip após expirar (não precisa mais de Procedure)
- Fixado não desconectar o OffAttack após  Vip Expirar
- Fixado por completo o Bug de Drop de bosses que as veses bugava e não dropava nada (na precisa mais do MonsterDieGiveItem.lua)
- Adicionado Custom Maps (Cliente/Lua/Configs/CustomMaps.lua) - (BETA)
- Adicionado Custom Capes (Cliente/Lua/ItemManager/CharacterCreateCape.lua) - (BETA)
- Adicionado Custom Pets com suporte a Miniaturas (Cliente/Lua/ItemManager/CharacterHelper.lua) - (BETA)
- Inserido AntiHacker Interno, apenas Habilite no MainInfo, e ligue o MHPServer no MuServer que já irá usa-lo sem precisar de DLL externa dificultando um pouco mais o Bypass
- Novo Sistema para usar SelectServers Customs, habilite no MainInfo caso vá usar algum selectserver que não possui movimento de camera (tour pelo mapa).
- Novo sistema para simular Bug Visual de HP do Ferrarezy fixado em 65k de vida

+++ Source/GameServer/GameServer/Monster.h
@@ -12,7 +12,7 @@ void gObjMonsterInitHitDamage(LPOBJ lpObj);
 void gObjMonsterSetHitDamage(LPOBJ lpObj,int aIndex,int damage);
 int gObjMonsterDelHitDamageUser(LPOBJ lpObj);
 int gObjMonsterGetTopHitDamageUser(LPOBJ lpObj);
-int gObjMonsterGetTopHitDamageParty(LPOBJ lpObj,int PartyNumber,int* TopHitDamageUser);
+QWORD gObjMonsterGetTopHitDamageParty(LPOBJ lpObj,int PartyNumber,int* TopHitDamageUser);
 int gObjAddMonster(int map);
 int gObjAddSummon();
 int gObjSummonSetEnemy(LPOBJ lpObj,int aIndex);
